# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {},
        "version": "2.0"
    },
    "basePath": "/api",
    "paths": {
        "/v1/cluster": {
            "get": {
                "description": "list cluster",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "k8s cluster"
                ],
                "summary": "list cluster",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "install cluster master",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "k8s cluster"
                ],
                "summary": "install cluster",
                "parameters": [
                    {
                        "description": "install cluster",
                        "name": "cluster",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cluster.clusterParam"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/v1/cluster/masters": {
            "post": {
                "description": "install cluster master",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "install cluster master"
                ],
                "summary": "install cluster",
                "parameters": [
                    {
                        "description": "install cluster master",
                        "name": "cluster",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Cluster"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/v1/cluster/slaves": {
            "post": {
                "description": "install cluster slave",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "k8s cluster"
                ],
                "summary": "install cluster slave",
                "parameters": [
                    {
                        "description": "install cluster slave",
                        "name": "cluster",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cluster.InstallSlave"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/v1/cluster/upgrade": {
            "post": {
                "description": "install cluster slave",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "k8s cluster"
                ],
                "summary": "upgrade k8s",
                "parameters": [
                    {
                        "description": "k8s all nodes",
                        "name": "upgrade",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/kubernetes.Upgrade"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "app.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "eventCode": {
                    "type": "integer"
                },
                "resMsg": {
                    "type": "object"
                }
            }
        },
        "cluster.Host": {
            "type": "object",
            "properties": {
                "ip": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "port": {
                    "type": "integer",
                    "default": 22
                },
                "user": {
                    "type": "string",
                    "default": "root"
                }
            }
        },
        "cluster.InstallSlave": {
            "type": "object",
            "properties": {
                "dryRun": {
                    "type": "boolean"
                },
                "master": {
                    "type": "object",
                    "$ref": "#/definitions/models.Host"
                },
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Host"
                    }
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "cluster.clusterParam": {
            "type": "object",
            "required": [
                "controlPlaneEndpoint",
                "master",
                "name"
            ],
            "properties": {
                "controlPlaneEndpoint": {
                    "type": "string"
                },
                "dryRun": {
                    "type": "boolean"
                },
                "master": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cluster.Host"
                    }
                },
                "name": {
                    "type": "string",
                    "default": "test"
                },
                "networkPlug": {
                    "type": "string",
                    "default": "cilium"
                },
                "podCidr": {
                    "type": "string",
                    "default": "10.244.0.0/16"
                },
                "registry": {
                    "description": "registry.aliyuncs.com/google_containers， k8s.gcr.io",
                    "type": "string",
                    "default": "registry.aliyuncs.com/google_containers"
                },
                "serviceCidr": {
                    "type": "string",
                    "default": "10.96.0.0/16"
                },
                "slaveNode": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cluster.Host"
                    }
                },
                "version": {
                    "type": "string",
                    "default": "1.22.15"
                }
            }
        },
        "kubernetes.Upgrade": {
            "type": "object",
            "properties": {
                "dryRun": {
                    "type": "boolean"
                },
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Host"
                    }
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "models.Cluster": {
            "type": "object",
            "required": [
                "controlPlaneEndpoint",
                "master",
                "name"
            ],
            "properties": {
                "controlPlaneEndpoint": {
                    "type": "string"
                },
                "master": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Host"
                    }
                },
                "name": {
                    "type": "string",
                    "default": "test"
                },
                "networkPlug": {
                    "type": "string",
                    "default": "cilium"
                },
                "podCidr": {
                    "type": "string",
                    "default": "10.244.0.0/16"
                },
                "registry": {
                    "description": "registry.aliyuncs.com/google_containers， k8s.gcr.io",
                    "type": "string",
                    "default": "registry.aliyuncs.com/google_containers"
                },
                "serviceCidr": {
                    "type": "string",
                    "default": "10.96.0.0/16"
                },
                "slaveNode": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Host"
                    }
                },
                "version": {
                    "type": "string",
                    "default": "1.22.15"
                }
            }
        },
        "models.Host": {
            "type": "object",
            "properties": {
                "ip": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "port": {
                    "type": "integer",
                    "default": 22
                },
                "user": {
                    "type": "string",
                    "default": "root"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        }
    }
}